+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.0](#030)
   - [0.2.1](#021)
   - [0.2.0](#020)
   - [0.1.6](#016)
@@ -11,6 +12,15 @@
   - [0.1.0](#010)
 
 ---
+
+## 0.3.0
+
+Released on 09/07/2024
+
+- Fix: resolved_host from configuration wasn't used to connect
+- `SshOpts::method` now requires `KeyMethod` and `MethodType` to setup key method
+- Feat: Implemented `SshAgentIdentity` to specify the ssh agent configuration to be used to authenticate.
+  - use `SshOpts.ssh_agent_identity()` to set the option
 
 ## 0.2.1